Fix review findings: move ChecksumCalculator to openfasttrack.importer
