+++ src/main/java/openfasttrack/importer/ChecksumCalculator.java
@@ -1,4 +1,4 @@
-package openfasttrack.core;
+package openfasttrack.importer;
 
 /*-
  * #%L
